backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.96.1
google-generativeai==0.8.5
python-docx==1.1.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2